+++ bibigrid-main/src/main/java/de/unibi/cebitec/bibigrid/StartUp.java
@@ -223,17 +223,24 @@ public class StartUp {
 
             // In order to validate the native instance types, we need a client.
             // So this step is deferred after client connection is established.
-            if (!validator.validateProviderTypes(client) || !validator.validateConfiguration()) {
-                LOG.error(ABORT_WITH_NOTHING_STARTED);
+            try {
+                if (!validator.validateProviderTypes(client) || !validator.validateConfiguration()) {
+                    LOG.error(ABORT_WITH_NOTHING_STARTED);
+                    return;
+                }
+            } catch (Exception e){
+                LOG.error(e.getMessage());
+                if (Configuration.DEBUG) {
+                    e.printStackTrace();
+                }
                 return;
             }
-
             switch (intentMode) {
                 case VALIDATE:
                     if (module.getValidateIntent(client, config).validate()) {
-                       LOG.info(I, "You can now start your cluster.");
+                        LOG.info(I, "You can now start your cluster.");
                     } else {
-                       LOG.error("There were one or more errors. Please adjust your configuration.");
+                        LOG.error("There were one or more errors. Please adjust your configuration.");
                     }
                     break;
                 case CREATE: