+++ ChangeLog.md
@@ -1,3 +1,12 @@
+## Version 2.2.2 (11/7/2020)
+
+### Fixes
+- Print possible configuration errors also to STDOUT (#222)
+- Add option to overwrite service CIDR mask (#283)
+
+### Features
+- Add Support for cloud site regions (#281)
+
 ## Version 2.2.1 (10/20/2020)
 
 ### Fixes